Use billablePercentage instead of displayedPercentage in certain places
In these cases it makes more sense to use billablePercentage. Before
this change, if one edits a cell to 100%, reloads, they are no longer
able to edit that cell and adjust it down. This also adds highlighting
for manually edited values (this previously only worked for cells edited
in the current session)